requirements.txt
py-cord==2.5.0
python-dotenv==1.0.1
mysql-connector-python
